Change build_index to only occur for old index
In current implementation a new index file will be created if the index file modification time is older than, or equal, to the fasta file. In some cases, e.g. if the files are downloaded as part of an external bundle, the modification times could be equal for all files. In such cases you often want to preserve the original index.

Fix this by only building index if the index is older than the fasta file. If they have equal mtime, it will be used instead.