fix: check for file before closing in __exit__
On error (e.g. opening non-existant files), the Faidx class __del__ is
called, which in turn calls  __exit__, which unconditionally does
self.file.close(). However if the file open or other such error casued
the file not to be opened, you get an attribute error in addition to the
original error. Therefore, the file.close() should be guarded behind
hasattr().